testing out cmake and moved headers
