+++ .gitignore
@@ -2,4 +2,4 @@
 .DS_Store
 journal.md
 website
-src_dummy
+src_reactor/build
