--- src_reactor/Makefile
@@ -1,36 +1,0 @@
-# Compiler and flags
-CXX = g++
-CXXFLAGS = -std=c++11 -Wall
-
-# Executable name
-TARGET = server
-
-# Source files
-SRCS = $(wildcard *.cpp)
-
-# Object files
-OBJS = $(SRCS:.cpp=.o)
-
-# Dependency files
-DEPS = $(SRCS:.cpp=.d)
-
-# Compile and link
-$(TARGET): $(OBJS)
-	$(CXX) $(CXXFLAGS) -o $(TARGET) $(OBJS)
-
-# Compile source files to object files
-%.o: %.cpp
-	$(CXX) $(CXXFLAGS) -c $< -o $@
-
-# Generate dependency files
-%.d: %.cpp
-	@$(CXX) $(CXXFLAGS) -MM $< | sed 's/$*\.o/& $@/g' > $@
-
-# Include dependency files
--include $(DEPS)
-
-.PHONY: clean
-
-# Clean up
-clean:
-	rm -f $(TARGET) $(OBJS) $(DEPS)
